+++ photon/4.0/inspec/vmware-photon-4.0-stig-baseline/controls/PHTN-40-000038.rb
@@ -1,5 +1,5 @@
 control 'PHTN-40-000038' do
-  title 'The Photon operating system must require the change of at 8 characters when passwords are changed.'
+  title 'The Photon operating system must require the change of at least 8 characters when passwords are changed.'
   desc  "
      If the operating system allows the user to consecutively reuse extensive portions of passwords, this increases the chances of password compromise by increasing the window of opportunity for attempts at guessing and brute-force attacks.
 
@@ -19,7 +19,7 @@ control 'PHTN-40-000038' do
 
     If the \"difok\" option is not >= 8, is missing or commented out, this is a finding.
 
-    Note: If pwquality.conf is not used to configure pam_pwquality.so then these options may be specified on the pwquality line in system-password file.
+    Note: If pwquality.conf is not used to configure pam_pwquality.so, these options may be specified on the pwquality line in the system-password file.
   "
   desc 'fix', "
     Navigate to and open:
@@ -44,9 +44,8 @@ control 'PHTN-40-000038' do
       its('difok') { should cmp >= 8 }
     end
   else
-    describe pam('/etc/pam.d/system-password') do
-      its('lines') { should match_pam_rule('password required pam_pwquality.so') }
-      its('lines') { should match_pam_rule('password required pam_pwquality.so').all_with_integer_arg('difok', '>=', 8) }
+    describe file('/etc/pam.d/system-password') do
+      its('content') { should match /^password\s+(required|requisite)\s+pam_pwquality\.so\s+(?=.*\bdifok=8\b).*$/ }
     end
   end
 end