+++ photon/4.0/inspec/vmware-photon-4.0-stig-baseline/controls/PHTN-40-000086.rb
@@ -19,7 +19,7 @@ control 'PHTN-40-000086' do
 
     If the \"ocredit\" option is not < 0, is missing or commented out, this is a finding.
 
-    Note: If pwquality.conf is not used to configure pam_pwquality.so then these options may be specified on the pwquality line in system-password file.
+    Note: If pwquality.conf is not used to configure pam_pwquality.so, these options may be specified on the pwquality line in the system-password file.
   "
   desc 'fix', "
     Navigate to and open:
@@ -44,9 +44,8 @@ control 'PHTN-40-000086' do
       its('ocredit') { should cmp < 0 }
     end
   else
-    describe pam('/etc/pam.d/system-password') do
-      its('lines') { should match_pam_rule('password required pam_pwquality.so') }
-      its('lines') { should match_pam_rule('password required pam_pwquality.so').all_with_integer_arg('ocredit', '<', 0) }
+    describe file('/etc/pam.d/system-password') do
+      its('content') { should match /^password\s+(required|requisite)\s+pam_pwquality\.so\s+(?=.*\bocredit=-1\b).*$/ }
     end
   end
 end