+++ photon/4.0/inspec/vmware-photon-4.0-stig-baseline/controls/PHTN-40-000206.rb
@@ -36,8 +36,7 @@ control 'PHTN-40-000206' do
   tag cci: ['CCI-000366']
   tag nist: ['CM-6 b']
 
-  describe pam('/etc/pam.d/system-auth') do
-    its('lines') { should match_pam_rule('auth optional pam_faildelay.so') }
-    its('lines') { should match_pam_rule('auth optional pam_faildelay.so').all_with_integer_arg('delay', '>=', 4000000) }
+  describe file('/etc/pam.d/system-auth') do
+    its('content') { should match /^auth\s+(required|requisite|optional)\s+pam_faildelay\.so\s+(?=.*\bdelay=4000000\b).*$/ }
   end
 end