+++ photon/4.0/inspec/vmware-photon-4.0-stig-baseline/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## [4.0 Version 1 Release 4] (2023-08-08)
+
+#### Release Notes
+- Updated source SRG to GPOS V2R6
+- PHTN-40-000105,111 updated severity to match SRG
+- PHTN-40-000005,47,49,59,79,130,160,193,199,239 misc. tech editing
+- PHTN-40-000012 updated check command and test
+- PHTN-40-000030 updated test to work when ran as non-root user
+- PHTN-40-000035,36,37,38,44,86,184,235 updated tests for use cases when not using pwquality.conf file and updated note in check text
+- PHTN-40-000043,243 updated tests for use cases when not using pwhistory.conf file. Removed order requirement for pam_pwhistory.so.
+- PHTN-40-000059 updated test
+- PHTN-40-000107 updated check to make it easier to assess this requirement
+- PHTN-40-000192 updated test
+- PHTN-40-000196 updated finding statements
+- PHTN-40-000197 updated test
+- PHTN-40-000206 updated test
+
 ## [4.0 Version 1 Release 3] (2023-06-06)
 
 #### Release Notes